novo commit de funções:

+++ src/main/java/CALISTO/model/dao/ContaDao.java
@@ -4,17 +4,18 @@ package CALISTO.model.dao;
 import CALISTO.model.persistence.Conta.Conta;
 import CALISTO.model.persistence.util.Conexao;
 
+import java.math.BigDecimal;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
 
 public abstract class ContaDao {
 
-    public boolean depositar(int idConta, double valor) {
+    public boolean depositar(int idConta, BigDecimal valor) {
         String sql = "UPDATE conta SET saldo = saldo + ? WHERE id_conta = ?";
         try (Connection conn = Conexao.getConnection();
              PreparedStatement stmt = conn.prepareStatement(sql)) {
-            stmt.setDouble(1, valor);
+            stmt.setBigDecimal(1, valor);
             stmt.setInt(2, idConta);
             return stmt.executeUpdate() > 0;
         } catch (SQLException e) {
@@ -23,13 +24,13 @@ public abstract class ContaDao {
         return false;
     }
 
-    public boolean sacar(int idConta, double valor) {
+    public boolean sacar(int idConta, BigDecimal valor) {
         String sql = "UPDATE conta SET saldo = saldo - ? WHERE id_conta = ? AND saldo >= ?";
         try (Connection conn = Conexao.getConnection();
              PreparedStatement stmt = conn.prepareStatement(sql)) {
-            stmt.setDouble(1, valor);
+            stmt.setBigDecimal(1, valor);
             stmt.setInt(2, idConta);
-            stmt.setDouble(3, valor);
+            stmt.setBigDecimal(3, valor);
             return stmt.executeUpdate() > 0;
         } catch (SQLException e) {
             e.printStackTrace();
@@ -37,7 +38,7 @@ public abstract class ContaDao {
         return false;
     }
 
-    public boolean transferir(int origemId, int destinoId, double valor) {
+    public boolean transferir(int origemId, int destinoId, BigDecimal valor) {
         Connection conn = null;
         try {
             conn = Conexao.getConnection();
